requirements.txt
python-docx==1.1.2
PyYAML==6.0.2